requirements.txt
llama-index==0.12.17
llama-index-llms-ollama==0.5.2
llama-index-embeddings-huggingface==0.5.1
python-dotenv==1.0.0
sentence-transformers==3.4.1
pypdf==5.3.0